src/django/requirements.txt
django-bootstrap3==9.1.0
django-cors-headers==2.2.0
djangorestframework==3.7.7
djangorestframework-gis==0.12.0
django-amazon-ses==0.3.0
django-annoying==0.10.4
django-bitfield==1.9.3
django-dbes==0.1.0
django-debug-toolbar==1.9.1
django-extensions==2.0.0
django-jsoneditor==0.1.4
django-premailer==0.2.0
django-queryset-csv==1.0.0
django-registration==2.4.1
django-rest-auth==0.9.3
django-watchman>=0.12,<0.12.99
factory-boy==2.10.0
flake8==3.5.0
ipdb==0.11
ipython==6.2.1
python-omgeo==5.0.0
requests>=2.21,<2.21.99
rollbar>=0.13,<0.13.99
us==1.0.0
